requirements.txt
SpeechRecognition>=3.14.0
PyAudio>=0.2.14
requests>=2.28.0
pyttsx3>=2.90
gtts>=2.5.0
pygame>=2.6.0

# Whisper ASR 相关依赖
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.25.0
numpy>=1.24.0
ffmpeg-python>=0.2.0 